Fix draw_dblpixel_raw when drawing on both layers. Fixes for Digic 6/7 version of zebra:   - remove redundant draw_guard_pixel function   - erase screen for blink modes when OSD restore option not set   - disable zebra if viewport larger then bitmap to prevent crash (e.g. using HDMI output) Disable edge overlay if viewport larger then bitmap to prevent crash (e.g. using HDMI output) Change vid_get_bitmap_fb on G5X and G7X2 to align with other ports. Remove unused label in G7X2 boot code.

+++ core/gui_draw.c
@@ -169,9 +169,9 @@ void draw_dblpixel_raw(unsigned int offset, unsigned int px, unsigned int op)
     bmbuf[offset] = px;
     opbuf[offset] = op | (op<<8);
 #else
-    ((unsigned short*)bitmap_buffer[0])[offset] = px;
+    ((unsigned int*)bitmap_buffer[0])[offset] = px;
     ((unsigned short*)opacity_buffer[0])[offset] = op | (op<<8);
-    ((unsigned short*)bitmap_buffer[1])[offset] = px;
+    ((unsigned int*)bitmap_buffer[1])[offset] = px;
     ((unsigned short*)opacity_buffer[1])[offset] = op | (op<<8);
 #endif
 }
